PR review fixing for migrations

+++ database/migrations/2020_09_23_151102_create_organization_user_roles_table.php
@@ -19,7 +19,7 @@ class CreateOrganizationUserRolesTable extends Migration
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
             $table->unique(['organization_id', 'user_id']);
-            $table->unsignedBigInteger('organization_role_type_id');
+            $table->unsignedInteger('organization_role_type_id');
             $table->foreign('organization_role_type_id')->references('id')->on('organization_role_types');
             $table->timestamps();
             $table->softDeletes();
