+++ database/migrations/2021_03_19_163531_create_organization_role_permissions_table.php
@@ -16,7 +16,7 @@ class CreateOrganizationRolePermissionsTable extends Migration
         Schema::create('organization_role_permissions', function (Blueprint $table) {
             $table->unsignedBigInteger('organization_role_type_id');
             $table->foreign('organization_role_type_id')->references('id')->on('organization_role_types');
-            $table->unsignedBigInteger('organization_permission_type_id');
+            $table->unsignedInteger('organization_permission_type_id');
             $table->foreign('organization_permission_type_id')->references('id')->on('organization_permission_types');
             $table->timestamps();
             $table->softDeletes();
